requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.0

# Claude API
anthropic==0.40.0

# Pinecone
pinecone-client==3.0.0

# Google Sheets
gspread==5.12.0
oauth2client==4.1.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.0

# Production Server
gunicorn==21.2.0
